refactor(app): move db connection and migration to app
