+++ kv/src/app/mod.rs
@@ -1,4 +1,7 @@
+use anyhow::Context;
 use sea_orm::*;
+
+use crate::{database, migrations};
 
 mod json;
 
@@ -8,8 +11,25 @@ pub struct App {
 }
 
 impl App {
-    pub fn new(db: DatabaseConnection) -> Self {
-        Self { db }
+    pub async fn new() -> anyhow::Result<Self> {
+        env_logger::init();
+
+        let data_dir = dirs::data_dir().context("getting data directory")?;
+        let db_dir = data_dir.join("kv");
+        std::fs::create_dir_all(&db_dir)?;
+        let db_path = db_dir.join("db");
+        std::fs::File::options()
+            .create(true)
+            .truncate(false)
+            .append(true)
+            .open(&db_path)
+            .expect("yes");
+        let db_url = format!("sqlite://{}", db_path.display());
+
+        let db = database::new(db_url).await?;
+        migrations::run(&db).await?;
+
+        Ok(Self { db })
     }
 }
