+++ kv/src/cli.rs
@@ -1,7 +1,6 @@
-use anyhow::Context;
 use clap::{Parser, Subcommand};
 
-use crate::{app::App, database, migrations};
+use crate::app::App;
 
 #[derive(Parser, Debug)]
 pub struct Cli {
@@ -27,24 +26,7 @@ pub enum JsonCommand {
 
 impl Cli {
     pub async fn run(self) -> anyhow::Result<()> {
-        env_logger::init();
-
-        let data_dir = dirs::data_dir().context("getting data directory")?;
-        let db_dir = data_dir.join("kv");
-        std::fs::create_dir_all(&db_dir)?;
-        let db_path = db_dir.join("db");
-        std::fs::File::options()
-            .create(true)
-            .truncate(false)
-            .append(true)
-            .open(&db_path)
-            .expect("yes");
-        let db_url = format!("sqlite://{}", db_path.display());
-
-        let db = database::new(db_url).await?;
-        migrations::run(&db).await?;
-
-        let app = App::new(db);
+        let app = App::new().await?;
 
         match self.command {
             Command::Json(json_command) => match json_command {
